Add diagnostic: dry deposition to subgrid-scale island
This adds a global dry deposition diagnostic (resolved by
species) which is supposed to represent the dry
deposition flux and dry deposition velocity to a subgrid-scale
sub-Antarctic island, much like Macquarie Is.

It works by temporarily replacing the modelled surface parameters
with ones which represent an ice-free island:

```
  Olson land type: cold grassland (42) for the entire grid cell
  Leaf area index: 1.0 (an approx match to Yuan BNU MODIS LAI
                        for Falkland Is.)
```

The usual dry deposition calculations are then performed with modified
parameters, results saved to new variables, and the normal model
state restored.

To enable this diagnostic output, the run's HISTORY.rc needs to have
the 'DryDep' collection enabled and the DryDep collection needs to be
modified to include the new variables, in addition to the
standard DryDep output, like this:

```
  DryDep.fields:              'DryDepVel_?DRY?               ',
                              'DryDep_?DRY?                  ',
                              'DryDepIsland_?DRY?            ',
                              'DryDepVelIsland_?DRY?         ',
```

Variables in the netCDF file will include the species name, for
example `DryDepIsland_Hg0` and `DryDepVelIsland_Hg0`.

There is no feedback to the model from the island, so this
diagnostic is conceptually only valid for very small islands
or on the windward side of islands within the belt of persistent
westerly winds, 50S-60S.